Correlation doc and code improvement
