+++ correlation/correlation.go
@@ -3,7 +3,6 @@ package correlation
 
 import (
 	"context"
-	"net/http"
 
 	"github.com/google/uuid"
 )
@@ -19,36 +18,21 @@ type idContextKey struct{}
 
 var idKey = idContextKey{}
 
+// New correlation ID.
+func New() string {
+	return uuid.New().String()
+}
+
 // IDFromContext returns the correlation ID from the context.
 // If no ID is set a new one is generated.
 func IDFromContext(ctx context.Context) string {
 	if id, ok := ctx.Value(idKey).(string); ok {
 		return id
 	}
-	return uuid.New().String()
+	return New()
 }
 
 // ContextWithID sets a correlation ID to a context.
 func ContextWithID(ctx context.Context, correlationID string) context.Context {
 	return context.WithValue(ctx, idKey, correlationID)
-}
-
-func GetOrSetHeaderID(h http.Header) string {
-	cor, ok := h[HeaderID]
-	if !ok {
-		corID := uuid.New().String()
-		h.Set(HeaderID, corID)
-		return corID
-	}
-	if len(cor) == 0 {
-		corID := uuid.New().String()
-		h.Set(HeaderID, corID)
-		return corID
-	}
-	if cor[0] == "" {
-		corID := uuid.New().String()
-		h.Set(HeaderID, corID)
-		return corID
-	}
-	return cor[0]
 }
